@@ -2,7 +2,7 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  before_filter :configure_permitted_parameters, if: :devise_controller?
+  before_action :configure_permitted_parameters, if: :devise_controller?
 
   # Enable rack-mini-profiler for signed in admin
   before_action do
@@ -80,20 +80,16 @@ class ApplicationController < ActionController::Base
   # Authorize if a user is signed in and is admin before viewing a pge
   def authorize_admin
   	unless current_user and current_user.admin
-  		# flash[:alert] = "Unauthorized access"
   		redirect_to home_path
-  		return false
+      throw :abort
   	end
-    return true #return true if the user is admin
   end
 
   # Authorize if a user is signed in
   def authorize_signed_in
     unless current_user
-      flash[:alert] = "You have to be signed in to do that!"
-      redirect_to user_session_path
-      return false
+      redirect_to user_session_path, alert: "You have to be signed in to do that!"
+      throw :abort
     end
-    return true #return true if the user is signed in
   end
 end
