@@ -1,4 +1,4 @@
-class Group < ActiveRecord::Base
+class Group < ApplicationRecord
   has_many :users_groups
   has_many :users, :through => :users_groups
 
