@@ -1,4 +1,4 @@
-class Notification < ActiveRecord::Base
+class Notification < ApplicationRecord
   belongs_to :sender,:class_name => 'User', :foreign_key => 'sender_id'
   belongs_to :receiver,:class_name => 'User', :foreign_key => 'receiver_id'
 end
