@@ -1,4 +1,4 @@
-class Relationship < ActiveRecord::Base
+class Relationship < ApplicationRecord
   # Made using https://www.railstutorial.org/book/following_users
   belongs_to :follower, class_name: "User"
   belongs_to :followed, class_name: "User"
