@@ -1,4 +1,4 @@
-class RepeatException < ActiveRecord::Base
+class RepeatException < ApplicationRecord
 	belongs_to :user
 	has_and_belongs_to_many :events
 	has_and_belongs_to_many :categories
