@@ -13,7 +13,7 @@ module RubyGettingStarted
     # -- all .rb files in that directory are automatically loaded.
 
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
+    # Run "rails -D time" for a list of tasks for finding time zone names. Default is UTC.
     config.time_zone = "UTC"
     config.active_record.default_timezone = :utc
 
