@@ -4,7 +4,7 @@ class ApplicationControllerTest < ActionController::TestCase
   include Devise::Test::ControllerHelpers
 
   test "search route works" do
-    get :search_core, q: "v"
+    get :search_core, params: { q: "v" }
     assert_response :success
   end
 end
