@@ -5,14 +5,12 @@ class HomeControllerTest < ActionController::TestCase
 
   test "should get homepage when not signed in" do
     get :index
-    assert_template("home/home", "home should be rendered if not signed in")
-    assert_response :success
+    assert_select "#home-sl1", true, "home should be rendered if not signed in"
   end
 
   test "should get dashboard when signed in" do
     sign_in users(:viktor)
     get :index
-    assert_template("home/dashboard", "dashboard should be rendered if signed in")
-    assert_response :success
+    assert_select "#greeting", true, "dashboard should be rendered if signed in"
   end
 end
