@@ -1,8 +1,8 @@
 # To run all tests, in the project directory run the command:
-# bundle exec rake test
+# bundle exec rails test
 # ----------------------------------------
 # To run this test, in the project directory run the command:
-# bundle exec rake test test/controllers/users_controller_test.rb
+# bundle exec rails test test/controllers/users_controller_test.rb
 
 require 'test_helper'
 
@@ -19,17 +19,17 @@ class UsersControllerTest < ActionController::TestCase
 
   test "signed in user should be able to view other user" do
     sign_in @viktor
-    get :show, id_or_url: @norm.id
+    get :show, params: { id_or_url: @norm.id }
     assert_response :success
   end
 
   test "non-signed in user should be able to view user" do
-    get :show, id_or_url: @norm.id
+    get :show, params: { id_or_url: @norm.id }
     assert_response :success
   end
 
   test "user views can be accessed through their custom urls" do
-    get :show, id_or_url: @putin.custom_url
+    get :show, params: { id_or_url: @putin.custom_url }
     assert_response :success
   end
 
@@ -42,7 +42,7 @@ class UsersControllerTest < ActionController::TestCase
   end
 
   test "routes to a user's id should redirect to their custom url when present" do
-    get :show, id_or_url: @viktor.id
+    get :show, params: { id_or_url: @viktor.id }
     assert_redirected_to user_path(@viktor)
   end
 
@@ -60,28 +60,28 @@ class UsersControllerTest < ActionController::TestCase
 
   test "signed in users can view their own profile" do
     sign_in @viktor
-    get :show, id_or_url: @viktor.id
+    get :show, params: { id_or_url: @viktor.id }
     assert_select "#profile-info", false,
                   "Users viewing themself should not see 'X Follower(s) You Know'"
   end
 
   test "users can navigate to the schedule tab" do
-    get :show, id_or_url: @norm.id, page: "schedule"
+    get :show, params: { id_or_url: @norm.id, page: "schedule" }
     assert_response :success
   end
 
   test "users can navigate to the followers tab" do
-    get :show, id_or_url: @norm.id, page: "followers"
+    get :show, params: { id_or_url: @norm.id, page: "followers" }
     assert_response :success
   end
 
   test "users can navigate to the following tab" do
-    get :show, id_or_url: @norm.id, page: "following"
+    get :show, params: { id_or_url: @norm.id, page: "following" }
     assert_response :success
   end
 
   test "users can navigate to the activity tab" do
-    get :show, id_or_url: @norm.id, page: "activity"
+    get :show, params: { id_or_url: @norm.id, page: "activity" }
     assert_response :success
   end
 
@@ -92,27 +92,27 @@ class UsersControllerTest < ActionController::TestCase
   # end
 
   test "users will navigate to the schedule tab by default" do
-    get :show, id_or_url: @norm.id, page: ""
+    get :show, params: { id_or_url: @norm.id, page: "" }
     assert_response :success
   end
 
   test "trying to view users that do not exist should redirect to the 404 page" do
-    get :show, id_or_url: "01010101010101"
+    get :show, params: { id_or_url: "01010101010101" }
     assert_response :missing
   end
 
   test "can perform an empty search query" do
-    get :search, q: nil
+    get :search, params: { q: nil }
     assert_response :success, "Accepts empty search queries"
   end
 
   test "can perform a search query" do
-    get :search, q: "v"
+    get :search, params: { q: "v" }
     assert_response :success
   end
 
   test "can perform a `json` search query" do
-    get :search, q: "v", format: "json"
+    get :search, params: { q: "v" }, format: "json"
     assert_response :success
   end
 end
